chore: Fix gitignore for accidental binary commits.
The current .gitignore patterns for accidental commits of binaries like
influxd and influx are too aggressive. It's common for tooling to use
.gitignore to filter out files from listings.

These rules cause the entire cmd/* directory to be excluded before the
exclusion (!*) rules are applied. Per gitignore manfile:

    An optional prefix "!" which negates the pattern; any matching file
    excluded by a previous pattern will become included again. It is not
    possible to re-include a file if a parent directory of that file is
    excluded.

The result is that any files/packages inside cmd/* directories cannot be found
by tooling  that uses .gitignore for file-filtering.

This adjusts the most permissive of the rules to target project root
only. Files found at other locations in the tree will be ignored
correctly.