feat: added a testing default value to feature flags
This change enables the setting of a testing default value for each feature
flag. The TestingFlagger() can then be used in test harnesses instead of the
DefaultFlagger() and the testing values are supplid to the Enabled() functions.

If a testing default value is not specified for a flag, then the default value
is used.

With this change we can default feature flags to true, or supply some other
default value for non-bool feature flags without needing to explicitly set them
in individual tests. At the same time the flags remain in their "safe state"
for when the code reaches production. Intention is to use this in unit tests,
end-to-end tests in influxdb, and in pipeline tests.