chore: update goreleaser dependency
The goreleaser project relies on the `gocloud.dev` dependency which
relies on an azure library. The azure library recently was changed to
become compatible with modules and one of the inner packages became its
own package rather than part of the base module.

This caused a conflict with flux which used the more recent azure
library and the version of goreleaser in influxdb which used the older
version. This commit updates the goreleaser version to the latest which
uses the latest version of `gocloud.dev` which has resolved this
dependency change.