feat(starter): add starter service to run scripts on events
events here are SetupComplete and StartupComplte. This example
code is an hour of coding something together. Not at all a
final product. It is a strawman to understand the basics of what
events can do for OSS.

What's missing here, a whole lot. The following would need to be addressed:

  * event service, single place to publish events to, decoupling pub/sub
  * simpler and more extensive events, likely an evnet type and properties
    root lvl of event struct
  * there is more things like tests and the like to make this production ready