chore(tools): remove goreleaser from the tools.go file
Goreleaser was causing conflict issues with our dependencies. The
version we relied on used an old go module that we needed to update, but
the module had been split in a backwards incompatible way so we needed
to update goreleaser at the same time.

The update then upgraded our protobuf dependency which we don't want at
the moment.

Since goreleaser is used as a tool rather than a formal dependency,
resolving these conflicts isn't really needed. We also do not use the
version listed in the go.mod file as we do not use `go run` to execute
goreleaser so this removes that dependency from `go.mod` since it is
unused.