feat: add support for multi-block readers
This commit embeds an index within each block materialised from a TSM
index, which can be used later on to identify which TSM block reader
should be used to decode the block.

This essentially lets one coalesce blocks for the same measurement from
different files lazily - that is, you don't need to materialise them
until you're ready, and when you do want to materialise them you know
which file to read from.