fix(ui): give user all access token permission to read user
Users used to be given the permission `orgs/<orgID>/users`, but this
does not help when the user is trying to access the users api (as it is
not nested under orgs in the api). We have changed the
`orgs/<orgID>/users` to be `users/<userID>` so that the token has access
to the user that created it. This will allow users to access their
tokens via the api.