fix(cmd/influx): query cli should not explicitly request gzipped content
The `net/http` package default transport will request content gzipped
automatically if `DisableCompression` is set to false (which is the
default). If the transport does it automatically, it will decompress it
automatically which is what was the original intention of the code. But,
if it is requested explicitly, then go will not decompress the data
automatically.