feat(storage/reads): replace panic on unsupported type with an error (#19353)
This replaces the panic that happens when attempting to create a cursor
for an unsupported aggregate type for the mean aggregate. It instead
refactors the result set so the cursor is constructed as part of the
`Next()` call and false is returned when there is an error creating the
cursor. The cursor is then accessed with `Cursor()`. When `Next()`
returns false, an error can be accessed by using `Err()`.