refactor: Simplify cancellation code
Prior to this commit, we used a select block with a default label to
check if the result of ctx.Done() has a pending message.  If so, it sets
a boolean value and later checks if it is true to determine if the loop
should exit.

This patch simply checks the result of ctx.Err() against nil.  This
should be much faster and will avoid allocating the channel that is
lazily created when ctx.Done() is called.