fix: Clarify what takes org/bucket names and what takes IDs
Connects to influxdata/fusion#60

As far as I can tell, `/api/v2/write` uses `org` and `bucket` params
that can take org and bucket names, so just call those `org` and
`bucket` throughout and clarify in examples they can be any string.

However, `/api/v2/buckets` uses only `orgID` that must be the
organization's 16-digit hex ID, so clarify that case.

This means in examples where we create a bucket and push points into it,
we'll need both a hex ID and name for the organization.