fix: Make influx client's URL public
In the data generator, we're going to have potentially 2 different
clients for the data and for telemetry, and I'd like to be able to log
how they're configured, for example.

I don't see any reason for URL to
be private, because the user of the client passed in the URL and we
haven't changed anything about it.