fix: Allow longer lines than the default for bufio.Scanner
This commit adds a field to Batcher to configure the maximum allowed
line length. If this value is non-zero, the bufio.Scanner buffer
will be configured scan lines up to this length. If a line exceeds
this length an ErrLineToLong error will be returned.

Simplified TestBatcher_read tests and added a test case for this
new scenario and to ensure lines exceeding the default
bufio.MaxScanTokenSize are allowed.