chore(ui): skip dashboard index CRUD case
I have it on good authority this is going away (or is being updated to
not make the same assertions) in a upcoming PR. The way dashboards load
templates has changed and it no longer relies on needing to POST to the
templates endpoint. So the way this test is structured is void.