chore(ae): add more logging
tsdb.Engine.IsIdle and tsdb.Shard.Digest now take a boolean parameter, isLogged.
If this is true, the engine will use trace-logging and report which checks are
causing the engine not to be idle.  tsdb.Shard exposes a funcion LoggedIsIdle()
which calls tsdb.Engine.IsIdle(true).  This is only used in calls from the
anti-entropy service in Enterprise.