feat: Add the git-log-prs script
__Scenarios__

You may know the git sha of what's running in the staging cluster,
and you may know the sha of the latest built image is, but will upgrading to a new
version include the PRs I'm interested in?

Or alternatively: I noticed in the dashboard that IOx was working fine until
we rolled out a new version. Which PRs were included in this new rollout?

__Description__

Getting the answers to the scenarios above is surprisingly hard, because scanning
our git history is complicated by the fact that just about anybody uses a different
merging technique.

This script does the right magic to skim through all that cruft and get the answers.

__Demo__

```console
$ ./scripts/git-log-prs 8376983b74311df970339e106c62ce4038b20e5f..
1330
1336
$ ./scripts/git-log-prs 8376983b74311df970339e106c62ce4038b20e5f.. --titles
1330 feat: Make background task period configurable
1336 feat: Build a perf_image image for every commit in main
```