fix: Anti-Entropy loops endlessly with empty shard (#21310)
The anti-entropy service will loop trying to copy an empty shard to a
data node missing that shard.  This fix is one of two changes that
correctly create an empty shard on a new node. This fix will set the
LastModified date of an empty shard directory to the modification time
of that directory, instead of to the Unix epoch.

Fixes: https://github.com/influxdata/influxdb/issues/21273
(cherry picked from commit 7f300dc248d767ee1a96642480dfa1daa28f18ba)

fixes https://github.com/influxdata/influxdb/issues/21287