perf(query): remove extraneous sort after push downs
When an ascending sort by time is included after a push down, it is now
removed. The reason for this is because all of the push downs return
points already sorted so the sort is not needed.

This is most important for `ReadGroup` because the native flux version
of `group()` does not return the data sorted so it might seem to be
necessary to use a sort after the group if you still need sorted data.
The push down does return sorted data so we can now remove that unneeded
sort.