feat(storage/flux): support table.fill() within aggregate window pushdown
This supports the `table.fill()` function within an
aggregate window.

    import "experimental/table"

    from(...)
    |> range(...)
    |> window(every: 5m, createEmpty: true)
    |> max()
    |> table.fill()
    |> duplicate(column: "_stop", as: "_time")
    |> window(every: inf)

This will now trigger the aggregate window pushdown and it will ensure
that there is a null row for windows where there are no points.  In
contrast, the previous behavior was that selectors that used aggregate
window would have created windows stripped and therefore never
created.

This can be used by adding it to a call to `aggregateWindow` like so:

    |> aggregateWindow(every: 5m, fn: (tables=<-, column) => tables |> max() |> table.fill())

At some point, we will add `table.fill()` to the `aggregateWindow()`
in the standard library.  When this is done, the above example will
become unnecessary but it will also still do the same thing because
multiple `table.fill()` calls will be merged into each other.