feat: Preserve full debug infos in release builds
The `Cargo.toml` file already says:

```toml
[profile.release]
debug = true
```

but we were overriding it in the circleci release script to include only line number info,
probably to shave off some time from the total build time.

The ability to debug a production binary outweighs the few seconds or even a minute overhead it takes to
add the debug infos. If building with full debug infos take much much longer, next time we disable them
it would be nice to include such a dramatic fact in the comments next to where we disable debug infos :-)